uint16_t -> uint32_t

+++ QD-FFT/fft/make_cos_table.hpp
@@ -4,7 +4,7 @@
 
 #include "../qd/qd.hpp"
 
-void make_quater_cos_table(uint16_t n, qd cos_table[]) {
+void make_quater_cos_table(uint32_t n, qd cos_table[]) {
     init(cos_table[0], 1.0);
     if (n <= 2)
         return;
@@ -16,7 +16,7 @@ void make_quater_cos_table(uint16_t n, qd cos_table[]) {
         return;
     std::queue<int> que;
     qd h;
-    uint16_t i, j, k;
+    uint32_t i, j, k;
     que.push(n / 8);
     while (!que.empty()) {
         i = que.front();
@@ -37,10 +37,10 @@ void make_quater_cos_table(uint16_t n, qd cos_table[]) {
     }
 }
 
-void make_cos_table(uint16_t n, qd cos_table[]) {
+void make_cos_table(uint32_t n, qd cos_table[]) {
     qd qcos_table[n / 4 + 1];
     make_quater_cos_table(n, qcos_table);
-    for(int i = 0;i < n / 4;i++) {
+    for(uint32_t i = 0;i < n / 4;i++) {
         copy(qcos_table[i], cos_table[i]);
         copy(qcos_table[n / 4 - i], cos_table[i + n / 4]);
         minus(cos_table[i + n / 4]);
@@ -50,10 +50,10 @@ void make_cos_table(uint16_t n, qd cos_table[]) {
     }
 }
 
-void make_sin_table(uint16_t n, qd sin_table[]) {
+void make_sin_table(uint32_t n, qd sin_table[]) {
     qd qcos_table[n / 4 + 1];
     make_quater_cos_table(n, qcos_table);
-    for(int i = 0;i < n / 4;i++) {
+    for(uint32_t i = 0;i < n / 4;i++) {
         copy(qcos_table[n / 4 - i], sin_table[i]);
         copy(qcos_table[i], sin_table[i + n / 4]);
         copy(qcos_table[n / 4 - i], sin_table[i + n / 2]);
